refactor proxy configuration
This commit changes proxy setting behaviour - now it's compatbile with Kubespray approach.
Due to Ansible limitations it's not possible to condionally define/undefine variables on playbook import, so proxy related vars have been moved from dictionary to top level and are used in the same way as they are in Kubespray.
After this change proxy_env dictionary is created in newly added cluster-defaults role based on the group_vars top-level string variables to preserve compatibility with our existing Ansible code.